refactor: 상품 수정 (#16)
테이블은 JPA 자동 생성 이용
테스트 데이터 삽입만 data.sql 작성
회원 역할, 회원 상태 데이터 자동으로 저장해주는 자바 설정파일 비활성화
회원 상태, 회원 역할 auto_increment 설정
전체 조회 수정
단건 상세 조회 수정
test 파일 삭제

+++ src/main/java/com/picketlogia/picket/api/product/repository/ProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface ProductRepository extends JpaRepository<Product,Long> {
-    Optional<Product> findByIdx(Integer idx);
+    Optional<Product> findByIdx(Long idx);
 
 }
